+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pyvirgo"
-version = "0.2.0-alpha.1"
+version = "0.2.0-alpha.2"
 description = "A Python package for the Virgo language"
 authors = ["Jack Grahl <jack.grahl@gmail.com>"]
 license = "MIT"
