+++ virgo/parse.py
@@ -5,18 +5,7 @@ import ply.yacc as yacc
 
 from .graph import make_graph
 from .internal import Connection
-
-
-COMMENT_REGEX = re.compile(r"[/]{2}.*$")
-
-def _gen_preprocess(data):
-    for line in data.split("\n"):
-        yield COMMENT_REGEX.sub("", line)        
-    yield ""
-
-
-def preprocess(data):
-    return "\n".join(_gen_preprocess(data))
+from .preprocess import preprocess
 
 
 tokens = (
